#3: Organize unit tests in correct packages.
