+++ src/test/java/com/commercetools/sync/products/utils/productupdateactionutils/BuildChangeSlugUpdateActionTest.java
@@ -1,4 +1,4 @@
-package com.commercetools.sync.products.utils;
+package com.commercetools.sync.products.utils.productupdateactionutils;
 
 import io.sphere.sdk.commands.UpdateAction;
 import io.sphere.sdk.models.LocalizedString;
